Atualizando os carros para alugar

+++ Provas_Antigas/Prova_Antiga_C206_L3_2022.1/src/main/java/org/example/Locadora.java
@@ -82,7 +82,7 @@ public class Locadora {
             return carrosParaAlugar;
         }
         else{
-            porcentagem = (carrosParaAlugar/total)*100;
+            porcentagem = ((float) carrosParaAlugar /total)*100;
             return porcentagem;
         }
     }
